+++ src/oci/generative_ai_inference/models/message.py
@@ -20,16 +20,20 @@ class Message(object):
     ROLE_SYSTEM = "SYSTEM"
 
     #: A constant which can be used with the role property of a Message.
-    #: This constant has a value of "USER"
-    ROLE_USER = "USER"
-
-    #: A constant which can be used with the role property of a Message.
     #: This constant has a value of "ASSISTANT"
     ROLE_ASSISTANT = "ASSISTANT"
 
     #: A constant which can be used with the role property of a Message.
+    #: This constant has a value of "USER"
+    ROLE_USER = "USER"
+
+    #: A constant which can be used with the role property of a Message.
     #: This constant has a value of "TOOL"
     ROLE_TOOL = "TOOL"
+
+    #: A constant which can be used with the role property of a Message.
+    #: This constant has a value of "DEVELOPER"
+    ROLE_DEVELOPER = "DEVELOPER"
 
     def __init__(self, **kwargs):
         """
@@ -40,12 +44,13 @@ class Message(object):
         * :class:`~oci.generative_ai_inference.models.AssistantMessage`
         * :class:`~oci.generative_ai_inference.models.UserMessage`
         * :class:`~oci.generative_ai_inference.models.ToolMessage`
+        * :class:`~oci.generative_ai_inference.models.DeveloperMessage`
 
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
         :param role:
             The value to assign to the role property of this Message.
-            Allowed values for this property are: "SYSTEM", "USER", "ASSISTANT", "TOOL", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "SYSTEM", "ASSISTANT", "USER", "TOOL", "DEVELOPER", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type role: str
 
@@ -84,6 +89,9 @@ class Message(object):
 
         if type == 'TOOL':
             return 'ToolMessage'
+
+        if type == 'DEVELOPER':
+            return 'DeveloperMessage'
         else:
             return 'Message'
 
@@ -93,7 +101,7 @@ class Message(object):
         **[Required]** Gets the role of this Message.
         Indicates who is writing the current chat message.
 
-        Allowed values for this property are: "SYSTEM", "USER", "ASSISTANT", "TOOL", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "SYSTEM", "ASSISTANT", "USER", "TOOL", "DEVELOPER", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -112,7 +120,7 @@ class Message(object):
         :param role: The role of this Message.
         :type: str
         """
-        allowed_values = ["SYSTEM", "USER", "ASSISTANT", "TOOL"]
+        allowed_values = ["SYSTEM", "ASSISTANT", "USER", "TOOL", "DEVELOPER"]
         if not value_allowed_none_or_none_sentinel(role, allowed_values):
             role = 'UNKNOWN_ENUM_VALUE'
         self._role = role