+++ CHANGELOG.rst
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 ====================
+2.158.2 - 2025-08-19
+====================
+
+Added 
+----- 
+* Support for calling Oracle Cloud Infrastructure services in the eu-budapest-1 region 
+* Support for pluggable database snapshot for Oracle Exadata Database Service 
+* Support for new GenericChatRequest parameters in Generative AI inference service 
+* Support for Flexible CIDR for secondary IPs on a VNIC for Virtual Cloud Network Service   
+
+====================
 2.158.0 - 2025-08-12
 ====================